Move search bar dependency views and drawables to common-ui
