+++ common-ui/src/main/res/drawable/ic_close_24dp.xml
