+++ common-ui/src/main/res/drawable/omnibar_field_background.xml
